+++ app/Http/Controllers/ConvocationController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Repositories\ConvocationRepository;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ConvocationController extends Controller
 {
@@ -12,6 +13,7 @@ class ConvocationController extends Controller
 
     public function __construct()
     {
+        $this->middleware('auth:sanctum')->except('');
         $this->repo = new ConvocationRepository();
     }
 
@@ -34,8 +36,12 @@ class ConvocationController extends Controller
      */
     public function store(Request $request)
     {
-        $array = $request->all();
-        return $this->repo->store($array);
+        if (Auth::user()->can('store_convocations')) {
+
+            $array = $request->all();
+            return $this->repo->store($array);
+        }
+        return abort(403);
     }
 
     /**
@@ -46,7 +52,11 @@ class ConvocationController extends Controller
      */
     public function show($id)
     {
-        return $this->repo->show($id);
+        if (Auth::user()->can('view_convocations')) {
+
+            return $this->repo->show($id);
+        }
+        return abort(403);
     }
 
     /**
@@ -58,8 +68,12 @@ class ConvocationController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $array = $request->all();
-        return $this->repo->update($array, $id);
+        if (Auth::user()->can('udpate_convocations')) {
+
+            $array = $request->all();
+            return $this->repo->update($array, $id);
+        }
+        return abort(403);
     }
 
     /**
@@ -70,6 +84,16 @@ class ConvocationController extends Controller
      */
     public function destroy($id)
     {
-        return $this->repo->delete($id);
+        if (Auth::user()->can('delete_convocations')) {
+
+            return $this->repo->delete($id);
+        }
+        return abort(403);
+    }
+
+
+    public function lastByCategory(int $category)
+    {
+        return $this->repo->lastByCategory($category);
     }
 }
